+++ web/src/features/menu-bar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MenuBar'
